+++ internal/codegen/expr_test.go
@@ -4,7 +4,7 @@ import (
 	"strings"
 	"testing"
 
-	"fct.dev/internal/parser"
+	"github.com/F33D3R-Inc/fct/internal/parser"
 )
 
 func TestExprLowering(t *testing.T) {